Save dictionary of data
Instead of saving 5 files per run which contain all the needed
information, we now store a dictionary of files. In addition, main.py
now works with multiple indices. For example main.py ... --indices 0 1
36 will run all experiments with hp settings indices between 0 and 35.
The data then generated from each run will be saved appropriately in a
Python dictionary and pickled. Now, we will be storing a single Python
pickle file, and NOT multiple files per run per hp setting. This will
significantly cut down on the number of files created.

In addition, for each hp setting, the dictionary will store all
environment and agent information so that if we have all the runs for an
experiment, we will not forget anything about the agent or environment.

Because main.py no longer works with --index, the main_concurrent.py may
no longer work. Also, because of concurrency, we would end up creating
multiple dictionary files for each run, which would only make sense if
it were for multiple runs of the same hp settings.