Fix bugs on cedar
Fixed some bugs that caused the program not to run on cedar.

First, now when running the subprocess module to run main.py, we specify
the absolute path, as this was causing extensive problems.

Second, changed the creation of the data directory in multiprocessing.
Due to race conditions, some processess would crash when multiple
processes saw that no data directory existed, and they all tried to
create the data directory. Only the first process to create this
directory would continue, and the others would crash.